+++ lib/pages/events_pages/event_list_view.dart
@@ -1,66 +1,33 @@
 import 'package:event_follow/models/controllers/events_controller/events_controller.dart';
-import 'package:event_follow/models/repositories/events/events_api_request.dart';
 import 'package:event_follow/pages/events_pages/event_content_view.dart';
 import 'package:event_follow/pages/events_pages/event_empty_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../config/sort_filter_globals.dart';
 import 'event_card.dart';
 
 class EventListView extends HookWidget {
   const EventListView({
+    required this.scrollController,
+    required this.onRefresh,
     Key? key,
   }) : super(key: key);
 
+  final ScrollController scrollController;
+  final RefreshCallback onRefresh;
+
   @override
   Widget build(BuildContext context) {
-    final controller = useProvider(eventsProvider);
     final data =
         useProvider(eventsProvider.state.select((value) => value.data));
-    final meta =
-        useProvider(eventsProvider.state.select((value) => value.meta));
     final isLoading =
         useProvider(eventsProvider.state.select((value) => value.isLoading));
-    final sortFilterStateStore = useProvider(eventsConditionProvider).state;
-
     final _cardList =
         data.map((datum) => EventCard(datum.event, datum.extra)).toList();
 
-    Future<void> _onRefresh() async {
-      await controller.request(EventsApiRequest(
-          pageId: '1',
-          sort: sortFilterStateStore?.sortType.typeName,
-          time: sortFilterStateStore?.timeFilterType?.typeName,
-          friends: sortFilterStateStore?.friendFilterType?.typeName));
-    }
-
-    late final _scrollController = () {
-      final _scrollController = ScrollController();
-      _scrollController.addListener(() {
-        final maxScrollExtent = _scrollController.position.maxScrollExtent;
-        final currentPosition = _scrollController.position.pixels;
-        if (maxScrollExtent > 0 &&
-            (maxScrollExtent - 100.0) <= currentPosition) {
-          final currentPage = meta?.currentPage ?? 1;
-          final totalPages = meta?.totalPages ?? 1;
-          if (!isLoading && _hasNextPaging(currentPage, totalPages)) {
-            final nextPageId = (meta?.currentPage ?? 1) + 1;
-
-            controller.request(EventsApiRequest(
-                pageId: nextPageId.toString(),
-                sort: sortFilterStateStore?.sortType.typeName,
-                time: sortFilterStateStore?.timeFilterType?.typeName,
-                friends: sortFilterStateStore?.friendFilterType?.typeName));
-          }
-        }
-      });
-      return _scrollController;
-    }();
-
     if (!isLoading && data.isEmpty) {
-      return EventEmptyView(onRefresh: _onRefresh);
+      return EventEmptyView(onRefresh: onRefresh);
     }
 
     if (isLoading && data.isEmpty) {
@@ -70,12 +37,8 @@ class EventListView extends HookWidget {
     }
 
     return EventContentView(
-        onRefresh: _onRefresh,
-        scrollController: _scrollController,
+        onRefresh: onRefresh,
+        scrollController: scrollController,
         cardList: _cardList);
-  }
-
-  bool _hasNextPaging(int currentPage, int totalPages) {
-    return currentPage < totalPages;
   }
 }
