+++ lib/pages/events_pages/events_page.dart
@@ -2,6 +2,7 @@ import 'package:event_follow/config/sort_filter_globals.dart';
 import 'package:event_follow/models/controllers/events_controller/events_controller.dart';
 import 'package:event_follow/models/repositories/events/events_api_request.dart';
 import 'package:event_follow/pages/events_pages/event_drawer_header.dart';
+import 'package:event_follow/pages/events_pages/scroll_top_button.dart';
 import 'package:event_follow/pages/events_pages/sort_filter_button.dart';
 import 'package:event_follow/pages/events_pages/sort_filter_dialog.dart';
 import 'package:flutter/material.dart';
@@ -14,6 +15,42 @@ class EventsPage extends HookWidget {
   @override
   Widget build(BuildContext context) {
     final controller = useProvider(eventsProvider);
+    final meta =
+        useProvider(eventsProvider.state.select((value) => value.meta));
+    final isLoading =
+        useProvider(eventsProvider.state.select((value) => value.isLoading));
+    final sortFilterStateStore = useProvider(eventsConditionProvider).state;
+
+    late final _scrollController = () {
+      final _scrollController = ScrollController();
+      _scrollController.addListener(() {
+        final maxScrollExtent = _scrollController.position.maxScrollExtent;
+        final currentPosition = _scrollController.position.pixels;
+        if (maxScrollExtent > 0 &&
+            (maxScrollExtent - 100.0) <= currentPosition) {
+          final currentPage = meta?.currentPage ?? 1;
+          final totalPages = meta?.totalPages ?? 1;
+          if (!isLoading && _hasNextPaging(currentPage, totalPages)) {
+            final nextPageId = (meta?.currentPage ?? 1) + 1;
+
+            controller.request(EventsApiRequest(
+                pageId: nextPageId.toString(),
+                sort: sortFilterStateStore?.sortType.typeName,
+                time: sortFilterStateStore?.timeFilterType?.typeName,
+                friends: sortFilterStateStore?.friendFilterType?.typeName));
+          }
+        }
+      });
+      return _scrollController;
+    }();
+
+    Future<void> _onRefresh() async {
+      await controller.request(EventsApiRequest(
+          pageId: '1',
+          sort: sortFilterStateStore?.sortType.typeName,
+          time: sortFilterStateStore?.timeFilterType?.typeName,
+          friends: sortFilterStateStore?.friendFilterType?.typeName));
+    }
 
     return Scaffold(
       appBar: AppBar(
@@ -57,7 +94,17 @@ class EventsPage extends HookWidget {
         ],
       ),
       drawer: EventDrawerHeader(),
-      body: const EventListView(),
+      body: EventListView(
+        scrollController: _scrollController,
+        onRefresh: _onRefresh,
+      ),
+      floatingActionButton: ScrollTopButton(
+        scrollController: _scrollController,
+      ),
     );
+  }
+
+  bool _hasNextPaging(int currentPage, int totalPages) {
+    return currentPage < totalPages;
   }
 }
